Add files via upload
"Block Dodger: Jahangir's Adventure" is an exciting game where you have to dodge falling blocks and avoid getting hit! Play as Jahangir, move to avoid the blocks, and enjoy the thrilling challenge!

Watch the gameplay and follow the steps shown in the videos to master your skills in the game. Check out the playlist for various game strategies and enjoy the adventure!

Click here to view the gameplay videos on YouTube:

https://youtube.com/playlist?list=PLagoYA7rdKmHu4p2rM7y6GGN6UUe_6-10&si=Ux6BOEzNkWniJGHn